Add "migrate" as an operation in the VM lifecycle job
This requires that the "destination" parameter be specified with the
target host name (as in host.name-label). This is only currently
supported for pool_migrate (i.e. not cross-pool storage motion).

Additionally allow "destination" to be optionally specified for the
"start" operation - this uses the start_on API to start on the
named host.